Fully separate demo logic per screen
