--- sample/src/main/java/com/simprints/sample/SimFaceDestination.kt
@@ -1,8 +1,0 @@
-package com.simprints.sample
-
-sealed interface SimFaceDestination {
-    data object Main : SimFaceDestination
-
-    data object Camera : SimFaceDestination
-}
-
