--- sample/src/main/java/com/simprints/sample/ui/models/camera/SimFaceCameraActions.kt
@@ -1,8 +1,0 @@
-package com.simprints.sample.ui.models.camera
-
-data class SimFaceCameraActions(
-    val onCaptureFace1: () -> Unit,
-    val onCaptureFace2: () -> Unit,
-    val onCompareCaptured: () -> Unit,
-)
-
