+++ sample/src/main/java/com/simprints/sample/ui/screens/camera/SimFaceCameraDemoScreen.kt
@@ -1,7 +1,10 @@
-package com.simprints.sample.ui.screens
+package com.simprints.sample.ui.screens.camera
 
 import android.Manifest
+import android.app.Application
 import android.content.pm.PackageManager
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.BackHandler
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.layout.Arrangement
@@ -15,6 +18,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -27,22 +31,54 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.core.content.ContextCompat
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.simprints.sample.di.SimFaceCameraViewModelFactory
 import com.simprints.sample.ui.composables.CameraCaptureSection
 import com.simprints.sample.ui.composables.ComparisonResultCard
 import com.simprints.sample.ui.composables.DisplayFaceResult
-import com.simprints.sample.ui.models.camera.SimFaceCameraActions
-import com.simprints.sample.ui.models.camera.SimFaceCameraUiState
 import kotlinx.coroutines.launch
 
 @Composable
 fun SimFaceCameraDemoScreen(
-    uiState: SimFaceCameraUiState,
-    actions: SimFaceCameraActions,
     snackbarHostState: SnackbarHostState,
     modifier: Modifier = Modifier,
 ) {
     val context = LocalContext.current
+    val application = remember(context) { context.applicationContext as Application }
+    val viewModel = remember(application) {
+        ViewModelProvider(
+            context as ComponentActivity,
+            SimFaceCameraViewModelFactory(application),
+        )[SimFaceCameraViewModel::class.java]
+    }
+    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
+
+    var showCameraPreview by remember { mutableStateOf(false) }
     val snackbarScope = rememberCoroutineScope()
+
+    LaunchedEffect(viewModel) {
+        viewModel.showSnackBarEffect.collect { message ->
+            snackbarHostState.showSnackbar(message)
+        }
+    }
+
+    BackHandler(enabled = showCameraPreview) { showCameraPreview = false }
+
+    if (showCameraPreview) {
+        CameraPreviewScreen(
+            modifier = modifier,
+            onDetectFaces = viewModel::detectFacesForPreview,
+            isProcessing = uiState.isProcessing,
+            onImageCaptured = {
+                viewModel.processCapturedBitmap(it)
+                showCameraPreview = false
+            },
+            onDismiss = { showCameraPreview = false },
+        )
+        return
+    }
+
     var hasCameraPermission by remember {
         mutableStateOf(
             ContextCompat.checkSelfPermission(context, Manifest.permission.CAMERA) ==
@@ -80,9 +116,19 @@ fun SimFaceCameraDemoScreen(
             isBusy = uiState.isProcessing || uiState.isComparing,
             capturedImage1 = uiState.capturedImage1,
             capturedImage2 = uiState.capturedImage2,
-            onCaptureFace1 = { checkAndRequestCameraPermission(actions.onCaptureFace1) },
-            onCaptureFace2 = { checkAndRequestCameraPermission(actions.onCaptureFace2) },
-            onCompareCaptured = actions.onCompareCaptured,
+            onCaptureFace1 = {
+                checkAndRequestCameraPermission {
+                    viewModel.setCameraTarget(CameraTarget.FACE_1)
+                    showCameraPreview = true
+                }
+            },
+            onCaptureFace2 = {
+                checkAndRequestCameraPermission {
+                    viewModel.setCameraTarget(CameraTarget.FACE_2)
+                    showCameraPreview = true
+                }
+            },
+            onCompareCaptured = viewModel::compareCapturedFaces,
         )
 
         if (uiState.isProcessing || uiState.isComparing) {
