+++ sample/src/main/java/com/simprints/sample/ui/screens/root/DemoTab.kt
@@ -1,7 +1,6 @@
-package com.simprints.sample.ui.models
+package com.simprints.sample.ui.screens.root
 
 enum class DemoTab {
     CAMERA,
     TEST_IMAGES,
 }
-
